feat: add google longrunning operations stubs
