+++ .gitattributes
@@ -1,0 +1,2 @@
+generated_types/protos/google/ linguist-generated=true
+generated_types/protos/grpc/ linguist-generated=true
