+++ generated_types/Cargo.toml
@@ -10,7 +10,7 @@ futures = "0.3.1"
 prost = "0.7"
 prost-types = "0.7"
 tonic = "0.4"
-
+tracing = "0.1"
 google_types = { path = "../google_types" }
 
 [build-dependencies] # In alphabetical order
